# services/design-frames-service/openapi.yaml
openapi: 3.0.3
info:
  title: design-frames-service
  version: 0.2.0
  description: >
    Product-design-phase service: navigable HTML frames, per-flow approval,
    and the API/component/flag contract seam. Extracted from FuzeFront's
    design/frames/** pipeline — see README.md and docs/EXTRACTION.md.


    v0.2.0 introduces the Postgres-backed lifecycle tier (contract only; the
    Express/pg runtime lands in a later stream). Frame CONTENT stays in git
    (the owning product's design/frames/<feature>/); Postgres holds ONLY
    lifecycle/index metadata + refs, keyed by content-stamp. Entities use the
    product-local `fxdf` identifier namespace (see .fuze/manifest.json and
    docs/postgres-tier.md). The entire v0.1.0 /api/v1/features/** surface is
    unchanged and backward-compatible — the new tier projects it from the new
    model. See docs/postgres-tier.md for the two-tier model and migration path.
  x-changelog:
    - version: 0.2.0
      date: '2026-08-19'
      changes:
        - 'ADD /api/v1/projects CRUD (create mints fxdf_prj_*, additionalProperties:false, no client id) + GET /{id}/features.'
        - 'ADD optional projectId reference on POST /api/v1/features (reference, not identity).'
        - 'ADD GET /api/v1/features/{slug}/flows/{flowId}/approvals — append-only decision history (paginated).'
...
    - version: 0.3.0
      date: '2026-10-09'
      changes:
...
    - version: 0.1.0
      date: '2026-08-11'
      changes:
...
servers:
  - url: http://localhost:4400
security:
  - bearerAuth: []
    delegatedUser: []
  - bearerAuth: []
  - {}
paths:
...
  /api/v1/projects:
    get:
      summary: List projects (paginated)
      security: []
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
                  page: { $ref: '#/components/schemas/PageInfo' }
    post:
      summary: Create a project (server mints the id)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/projects/{id}:
    parameters:
      - { $ref: '#/components/parameters/ProjectIdPath' }
    get:
      summary: Get one project
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update a project's mutable fields
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/features:
    get:
      summary: List the features belonging to a project (paginated)
      security: []
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureSummary' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/flows:
    parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
    get:
      summary: List project-owned native UX flows
      security: []
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/NativeFlow' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
    post:
      summary: Create a native UX flow and its first immutable document revision
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NativeFlowCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [flow, revision]
                properties:
                  flow: { $ref: '#/components/schemas/NativeFlow' }
                  revision: { $ref: '#/components/schemas/FlowDocumentRevision' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/projects/{id}/flows/{flowId}:
    get:
      summary: Get one project-owned native UX flow
      security: []
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/FlowIdPath' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/NativeFlow' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/flows/{flowId}/revisions:
    parameters:
      - { $ref: '#/components/parameters/ProjectIdPath' }
      - { $ref: '#/components/parameters/FlowIdPath' }
    get:
      summary: List immutable native flow document revisions
      security: []
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/FlowDocumentRevision' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
    post:
      summary: Append a native flow document revision with optimistic concurrency
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlowDocumentRevisionCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/FlowDocumentRevision' } } } }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: expectedRevision differs from current latest revision }
  /api/v1/projects/{id}/flows/{flowId}/revisions/{revision}:
    get:
      summary: Read one immutable native flow document revision
      security: []
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/FlowIdPath' }
        - name: revision
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 2147483647 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FlowDocumentRevision' } } } }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/repositories:
    parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
    get:
      summary: List database-managed repository connections for an App workspace
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectRepository' }
    post:
      summary: Connect a repository to an App and adopt matching imported frames
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectRepositoryCreate' }
      responses:
        '201':
          description: Connected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectRepository' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/projects/{id}/workspace:
    get:
      summary: Get an app workspace with manifest-backed counts and its latest design system
      security: []
      parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectWorkspace' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/design-system:
    get:
      summary: Get the latest immutable design-system revision
      security: []
      parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignSystemRevision' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/design-system/revisions:
    parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
    get:
      summary: List design-system revisions in ascending revision order
      security: []
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/DesignSystemRevision' } }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Append a design-system snapshot with an optimistic predecessor check
      description: >
        expectedRevision must equal the current latest revision (0 initially).
        Component decisions and token edits create a new snapshot; old snapshots
        remain readable. createdBy is taken from the verified caller.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DesignSystemRevisionCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignSystemRevision' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: expectedRevision differs from the current latest revision }
  /api/v1/projects/{id}/design-system/revisions/{revision}:
    get:
      summary: Read a historical design-system snapshot
      security: []
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - name: revision
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 2147483647 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DesignSystemRevision' }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/trace-links:
    parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
    get:
      summary: List durable FuzePlan/FuzeQuality decision evidence for an App
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Attach immutable business or QA evidence to a UX target
      description: External URLs are display-only evidence locators and are never fetched by FuzeX.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TraceLinkCreate' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/TraceLink' } } } }
        '400': { $ref: '#/components/responses/Validation' }
  /api/v1/projects/{id}/design-policies:
    parameters: [{ $ref: '#/components/parameters/ProjectIdPath' }]
    get:
      summary: List scoped design-agent policies; approvedOnly=true is safe for agent consumption
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK } }
    post:
      summary: Draft a natural-language design policy linked to decision evidence
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DesignPolicyCreate' } } }
      responses: { '201': { description: Created } }
  /api/v1/projects/{id}/design-policies/{policyId}/approve:
    post:
      summary: Explicitly approve a draft policy for AI-agent consumption
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - name: policyId
          in: path
          required: true
          schema: { type: string, pattern: '^fxdf_dpl_[0-9a-hjkmnp-tv-z]+$' }
      responses:
        '200': { description: Approved }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/projects/{id}/design-policies/{policyId}/supersede:
    post:
      summary: Supersede an approved or draft policy with a new immutable draft
      description: The replacement is a new draft linked to the prior policy; the prior instruction is retained as historical evidence and removed from agent consumption.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - name: policyId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/DesignPolicyId' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DesignPolicyCreate' } } }
      responses:
        '201': { description: Replacement draft created }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    delegatedUser:
      type: apiKey
      in: header
      name: X-Fuze-Delegation
      description: >
        Server-to-server header with value Bearer <delegation-token>. FuzeFront's
        same-origin proxy exchanges the browser session for a fuze-delegation
        token with audience service:fuzex. Authorization must contain its
        immediate actor's fuze-workload token. Both are introspected; actor.sub
        must match the workload subject and tenant-bound workloads must match
        the delegated tenant. Delegation requires fuzex:frames:read for reads
        and fuzex:frames:write for mutations. Invalid supplied delegation never
        falls back to anonymous reads or direct machine writes. Browsers never
        receive either upstream credential.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A FuzeFront-issued MACHINE token for direct service writes, or the
        fuze-workload token accompanying delegatedUser for hosted user calls.
        Reads without delegation are public. All write methods require auth.

        Obtain one from FuzeFront's `POST /api/v1/security/tokens`. This service
        verifies it per request against FuzeFront's
        `POST /api/v1/security/tokens/introspect` and requires the scope
        `fuzex:frames:write`; a token without it gets 403, not 401.

        Verification is fail-closed: introspection answers HTTP 200 even for an
        inactive token, so the decision is made on the `active` boolean in the
        response body. An unreachable or unusable introspection endpoint denies
        the request rather than allowing it.

        This replaced the `DESIGN_FRAMES_API_TOKENS` pre-shared bearer list
        (issue #26), which is deleted — there is no fallback.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: 'Max items per page. Default 50, max 200.'
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      description: 'Opaque forward cursor from a prior page.nextCursor. Absent = first page.'
      schema: { type: string }
  responses:
...
    Unauthorized:
      description: Missing/invalid bearer token on a write operation
...
    Conflict:
      description: Slug/name already exists
...
  schemas:
...
    PageInfo:
      type: object
      required: [nextCursor, hasMore]
      description: 'Cursor pagination envelope page block (baseline §4.1).'
      properties:
        nextCursor:
          type: string
          nullable: true
          description: 'Opaque cursor for the next page, or null when hasMore is false.'
        hasMore: { type: boolean }
        total:
          type: integer
          nullable: true
          description: 'Total matching rows when cheaply known; otherwise omitted/null.'
...